WIP model checkpoint loading

+++ src/tsdm/utils/remote.py
@@ -18,11 +18,15 @@ from tsdm.types.aliases import PathLike
 
 
 def hash_file(
-    file: str, block_size: int = 65536, algorithm: str = "sha256", **kwargs: Any
+    file: PathLike,
+    *,
+    block_size: int = 65536,
+    hash_algorithm: str = "sha256",
+    **kwargs: Any,
 ) -> str:
     r"""Calculate the SHA256-hash of a file."""
     algorithms = vars(hashlib)
-    hash_value = algorithms[algorithm](**kwargs)
+    hash_value = algorithms[hash_algorithm](**kwargs)
 
     with open(file, "rb") as file_handle:
         for byte_block in iter(lambda: file_handle.read(block_size), b""):
@@ -31,13 +35,41 @@ def hash_file(
     return hash_value.hexdigest()
 
 
+def validate_hash(
+    fname: PathLike, hash_value: str, *, hash_algorithm: str = "sha256", **kwargs: Any
+) -> None:
+    r"""Validate a file against a hash value."""
+    hashed = hash_file(fname, hash_algorithm=hash_algorithm, **kwargs)
+    if hashed != hash_value:
+        raise ValueError(
+            f"Calculated hash value {hashed!r} and given hash value {hash_value!r} "
+            f"do not match for given file {fname!r} (using {hash_algorithm})."
+        )
+
+
 def download(
-    url: str, fname: Optional[PathLike] = None, *, chunk_size: int = 1024
+    url: str,
+    fname: Optional[PathLike] = None,
+    *,
+    chunk_size: int = 1024,
+    skip_existing: bool = True,
+    hash_value: Optional[str] = None,
+    hash_algorithm: str = "sha256",
+    hash_kwargs: Optional[dict[str, Any]] = None,
 ) -> None:
     r"""Download a file from a URL."""
+    hash_kwargs = {} if hash_kwargs is None else hash_kwargs
     response = requests.get(url, stream=True, timeout=10)
     total = int(response.headers.get("content-length", 0))
     path = Path(fname if fname is not None else url.split("/")[-1])
+
+    if skip_existing and path.exists():
+        if hash_value is not None:
+            validate_hash(
+                path, hash_value, hash_algorithm=hash_algorithm, **hash_kwargs
+            )
+        return
+
     try:
         with path.open("wb") as file, tqdm(
             desc=str(path),
@@ -55,10 +87,8 @@ def download(
             f"Error {e!r} occurred while downloading {fname}, deleting partial files."
         ) from e
 
-
-def validate_hash(fname: str, hash_value: str, hash_type: str = "sha256") -> bool:
-    r"""Validate a file against a hash value."""
-    return hash_file(fname, algorithm=hash_type) == hash_value
+    if hash_value is not None:
+        validate_hash(path, hash_value, hash_algorithm=hash_algorithm, **hash_kwargs)
 
 
 def to_base(n: int, b: int) -> list[int]:
